Add configure macro to prohibit compiler warnings and remove all warnings from sipXportLib.
To prohibit warnings, add:

 SFAC_STRICT_COMPILE_NO_WARNINGS_ALLOWED

to the configure.ac file (it must be before SFAC_INIT_FLAGS, but
autoconf will complain if you get that part wrong).  With this macro in
configure.ac, when gcc is in use the -Werror flag is added by default
and any warning stops the build.  This can be turned off using the
--disable-compile-strict argument to the 'configure' script, but the
automated builds wouldn't do that, so checking in code that generated a
warning in a 'strict' subproject would break its build.  There is no
effect on a subproject that does not use the macro.

http://list.sipfoundry.org/archive/sipx-dev/msg08236.html



git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10909 ab1d8caa-1f67-47f1-9e81-24633a41865c